refactor product folder structure
